requirements.txt
PyQt6>=6.6.0
PyQt6-tools>=6.6.0
psycopg2-binary>=2.9.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
